# app/src/Controller/Api/Auth/SignUpController.php
<?php
declare(strict_types=1);

namespace App\Controller\Api\Auth;

use App\Controller\ErrorHandler;
use App\Model\User\UseCase\SignUp;
use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
use Symfony\Component\HttpFoundation\JsonResponse;
use Symfony\Component\HttpFoundation\Request;
use Symfony\Component\HttpFoundation\Response;
use Symfony\Component\Routing\Annotation\Route;
use Symfony\Component\Serializer\SerializerInterface;
use Symfony\Component\Validator\Validator\ValidatorInterface;
use OpenApi\Annotations as OA;

class SignUpController extends AbstractController
{
    private $serializer;
    private $validator;
    private $errors;

    public function __construct(SerializerInterface $serializer, ValidatorInterface $validator, ErrorHandler $errors)
    {
        $this->serializer = $serializer;
        $this->validator = $validator;
        $this->errors = $errors;
    }

    /**
     *     @OA\Post(
     *     path="/auth/signup",
     *     tags={"Sign Up"},
     *     @OA\RequestBody(
     *         @OA\JsonContent(
     *             type="object",
     *             required={"first_name", "last_name", "email", "password"},
     *             @OA\Property(property="first_name", type="string"),
     *             @OA\Property(property="last_name", type="string"),
     *             @OA\Property(property="email", type="string"),
     *             @OA\Property(property="password", type="string"),
     *         ),
     *     ),
     *     @OA\Response(
     *         response=201,
     *         description="Success response",
     *     ),
     *     @OA\Response(
     *         response=400,
     *         description="Errors",
     *         @OA\JsonContent(
     *             type="object",
     *             @OA\Property(property="error", type="object", nullable=true,
     *                 @OA\Property(property="code", type="integer"),
     *                 @OA\Property(property="message", type="string"),
     *             ),
     *             @OA\Property(property="violations", type="array", nullable=true, @OA\Items(
     *                 type="object",
     *                 @OA\Property(property="propertyPath", type="string"),
     *                 @OA\Property(property="title", type="string"),
     *             ))
     *         )
     *     ),
     * )
     *
     * @Route("/auth/signup", name="auth.signup", methods={"POST"})
     * @param Request $request
     * @param SignUp\Request\Handler $handler
     * @return Response
     */
    public function request(Request $request, SignUp\Request\Handler $handler): Response
    {
        /** @var SignUp\Request\Command $command */
        $command = $this->serializer->deserialize($request->getContent(), SignUp\Request\Command::class, 'json');

        $violations = $this->validator->validate($command);
        if (\count($violations)) {
            $json = $this->serializer->serialize($violations, 'json');
            return new JsonResponse($json, 400, [], true);
        }


        return $this->json([], 201);
    }
}